Move auth controllers into base controller folder
