+++ app/controllers/facility_reviews_controller.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-class FacilityReviewsController < AuthenticateController
+class FacilityReviewsController < BaseControllers::AuthenticateController
   def new
     @categories = FacilityCategory.all
     @space = Space.find(params["space_id"])
