+++ app/controllers/spaces_controller.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-class SpacesController < AuthenticateController # rubocop:disable Metrics/ClassLength
+class SpacesController < BaseControllers::AuthenticateController # rubocop:disable Metrics/ClassLength
   def index
     @spaces = Space.all.order updated_at: :desc
     @space = Space.new
